demo/requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
grpcio==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hyperframe==6.0.1
idna==3.7
jinja2==3.1.3
joblib==1.4.0
jsonlines==4.0.0
llama-index==0.10.29
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.11
llama-index-core==0.10.29
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-ollama==0.1.2
llama-index-llms-openai==0.1.15
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-qdrant==0.2.0
llama-parse==0.4.1
llamaindex-py-client==0.1.18
markupsafe==2.1.5
marshmallow==3.21.1
minijinja==1.0.16
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.19.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
portalocker==2.8.2
protobuf==4.25.3
pydantic==2.7.0
pydantic-core==2.18.1
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
qdrant-client==1.8.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.6.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlalchemy==2.0.29
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
typing-extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
llama-index-experimental
ray
